requirements.txt
python-dotenv
cx-Oracle